Updated on 15 Jan 2025

+++ backend/app/api/controllers/tickets.js
@@ -22,7 +22,9 @@ const removeUploadedFiles = (files) => {
     }
     else {
         files.forEach((file) => {
-            fs.unlinkSync(file.path)
+            if(fs.existsSync(file.path)) {
+                fs.unlinkSync(file.path)
+            }
         })
     }
 }
